Disable E402 in sync-engine
Summary:
E402 is everywhere in the sync-engine repo and is causing builds
to be marked as unbuildable. Disable it. (Note: a similar commit has
already landed in sync-engine-eas. Pyflakes looks for tox.ini
on a project level, so each repo is going to need this added
if we want to disable the lint everywhere.)

Test Plan: This run of arc diff is the test.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://phab.nylas.com/D2563