Update IMAP/ SMTP credentials on re-authing a generic account.
Summary:
We absolutely must update the IMAP/ SMTP credentials in update_account()
in order for the legacy auth to function as desired.

Test Plan:
Regression test added.
Tested entire flow manually. I attempted to write an end-to-end test in redwood
to do so as well and am extremely unsatisfied with it. Let's not block on it though; I
will write it shortly.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://phab.nylas.com/D2606