Conditionally expose the `account` object in the delta, streaming APIs.
Summary:
Currently, no notification is sent to clients when an account becomes invalid.
This is problematic for N1 since there is no way to "gracefully fail" to the user and ask
them to reauthenticate until the next failing API request.

To address this, conditionally expose the `account` object in the delta, streaming APIs --
by setting the `exclude_account` flag to False, N1 will receive notifications about account.sync_state changes
(i.e. when it becomes invalid, when it flips to running again).

Test Plan: Tests added.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://phab.nylas.com/D2657