Return a consistent response across all API endpoints for an invalid account.
Summary:
As is, the different API endpoints return different and even internally inconsistent responses for invalid accounts:
* /send returns an HTTP 403 for Gmail, generic IMAP accounts,
* /search returns an HTTP 403 for Gmail accounts and an HTTP 500 for generic IMAP,
* More troublingly, message and thread update operations can return an HTTP 200 if there is no state change,
despite the account being invalid -- since the account validity check only occurs when the remote action is scheduled, and
an HTTP 403 if there is a state change.

To address all of the above, verify the validity of the account's credentials /before/ performing a
request to the remote server. This includes the message and thread /search endpoints, and the /send endpoint
which directly interact with the remote server, and all create, update, delete requests which result in requests
to the remote server via action syncback.

NOTE:
The concomitant EAS changes will be submitted in a separate revision.

Test Plan:
Tests added.
Tested manually.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://phab.nylas.com/D2658