Add utility script for unscheduling all syncs on a given host
Summary:
This is primarily useful for unscheduling all the accounts on a single sync
host after terminating it. This codifies something I did manually the other
week, and would like for people to not have to figure out how to do every time
we need it.

Test Plan: Tested manually with --dry-run and looked at output

Reviewers: ethanb

Reviewed By: ethanb

Differential Revision: https://phab.nylas.com/D2770