Added logging of Google Calendar endpoint & ID info
Summary:
In order to re-enable the /w/calendar_update endpoint, we want to log
more data about incoming Google Calendar requests so that we can see if we have
duplicate subscriptions to Google's notifications. If we are, we can fix the
problem most quickly by unsubscribing from the duplicates, although rate limits
are independently a good idea. This issue is tracked in T6714.

Test Plan:
I've confirmed that this doesn't break the ability to spin up a dev
instance, whereas a previous iteration with a syntax error did, so that's a
good first step. I am a bit unsure how to test this properly and would welcome
suggestions on that point.

Reviewers: kav-ya

Reviewed By: kav-ya

Subscribers: ethanb

Projects: #reliability

Maniphest Tasks: T6714

Differential Revision: https://phab.nylas.com/D2832