Handle IMAP folders correctly (1/2)
Summary:
D2503 changed the way we handle IMAP folder separators: we're now storing IMAP folder separator and prefixes per-account, instead of assuming all non-Gmail servers used `.`.

Unfortunately, I couldn't ship it to prod because it required running a lengthy backfix, which caused our API success rate to drop.

To be able to ship this diff without affecting our customers, I've decided to split it into two parts:
1. the first part will contain the backfix code and changes to the auth handler.
2. the second part will contain the rest of the diff.

I'm going to ship the first part to prod, run the backfix and, when it's done, ship the second part.

Test Plan: Tested on staging when it was D2503.

Reviewers: kav-ya

Reviewed By: kav-ya

Differential Revision: https://phab.nylas.com/D2739