Selectively overwrite sync_{host, state} while stopping syncs.
Summary:
Invoking queue_client.unassign() /before/ stopping an account's sync can have the
undesired effect of having the sync be reclaimed and started on a new host before the
original host removes it from its set of accounts to sync. When this happens, the account
may be syncing on more than one host for a period, and may be subsequently incorrectly set to
'stopped' when the original host eventually removes it.

The right solution to this is to update the unscheduling script to hit the `/unassign` endpoint exposed by
each mailsync process -- which will first stop the account's sync and then call queue_client.unassign() for it.

In the meantime, update account.sync_stopped() to perform a compare-and-swap before updating values --
only iff the requesting host still owns the account sync, account.sync_{host, state} are updated.
This is useful to do independant of updating the unscheduling script,  in case we ever need to unassign
using the `queue_client` directly.

Test Plan:
Regression test added that verifies:
* An incorrect host cannot update these values.
* The correct host can.

Reviewers: spang

Reviewed By: spang

Differential Revision: https://phab.nylas.com/D2910