Faster S3 sync
Summary:
This diff adds a new type of FolderSyncEngine name S3FolderSyncEngine. This engine is created by the ImapSyncMonitor.

The S3FolderSyncEngine is a dumbed down version of the foldersyncengine we use for syncing generic IMAP accounts --- it simply syncs every uid and uploads them to S3. To make sure that we're not downloading the same messages over and over, we save our progress in Account._sync_status. It's a JSON column which is only used by the status code; I felt it wasn't worth to write a migration just for this.

I still need to write a similar program for Exchange. It should be much easier because we could just use the primary device to get all the messages from the Exchange server.

Test Plan: Tested manually, will re-check my results and test on staging.

Reviewers: kav-ya, spang

Reviewed By: spang

Subscribers: jenkins

Maniphest Tasks: T6857

Differential Revision: https://phab.nylas.com/D2945