Don't unload accounts at every turn of the loop
Summary:
Right now, if the CPU usage of an machine is over a set threshold, we unload an account every 10 seconds. This is _very_ aggressive, especially since we run 16 mailsync process per machine.

This diff changes this to unloading an account every minute.

Test Plan: Wrote a regression test. Will test on staging.

Reviewers: spang

Reviewed By: spang

Subscribers: jenkins

Differential Revision: https://phab.nylas.com/D3063