Move multi-send 'delete from sent' to session close, run for gmail only
Summary:
Deleting copies of the sent message from the sent folder is only relevant
for gmail. Also, doing it as part of the individual send was causing
timeouts when many /send-multiple requests came in quickly (i.e. message
has many recipients). This moves the deletion to the final DELETE call that
also moves the draft to the sent folder. This requires only one call to the
remote provider (instead of one call per recipeint), and can be safely
retried if it fails.

Test Plan: Ensure current tests pass, test with many recipients on staging.

Reviewers: spang

Reviewed By: spang

Subscribers: jenkins

Differential Revision: https://phab.nylas.com/D3066