Send a cancellation message for users who have been removed from an event
Summary:
This diff fixes an oversight with the invite sending code --- if you updated an invite by removing an attendee, they wouldn't receive an event cancellation message, which means that the event would still show up in their calendar.

To fix this issue we get the list of attendees which got removed and save it into the Action's extra_args field (we make sure to truncate it to not get over the 64k chars limit of `TEXT` fields). Then, after executing the action we send a message to the attendees.

This works great from Exchange to Gmail and Exchange to Fastmail but cancelling an invite from Exchange to Exchange doesn't work great --- you receive the cancellation but for some reason Outlook won't cancel it. I'm still looking into this issue but it's definitely not a show stopper.

Test Plan: Tested manually. Will need more extensive testing.

Reviewers: drew, spang

Reviewed By: spang

Subscribers: jenkins, tomasz

Differential Revision: https://phab.nylas.com/D3083