Fix flaky transaction test by threading 'now' through
Summary:
We were seeing some flakiness on Jenkins due to an inconsistent use of 'now' when
creating and purging transactions. This diff threads the notion of our starting
'now' through the test and into the utility functions it depends on.

Test Plan: Run unit test, make sure it passes. Watch for future flakiness

Reviewers: spang

Reviewed By: spang

Subscribers: jenkins

Differential Revision: https://phab.nylas.com/D3089