[SCHEMA] Add queryable_value column to Metadata
Summary:
Adds a new integer column to the Metadata table to allow querying for
metadata by value. The column functions as in index on a particular value
in the metadata's JSON value. The column will be populated in application
code based on the Metadata.value field, using a specified field name per
app.

Test Plan: Will be tested along with dependent code in redwood.

Reviewers: spang

Reviewed By: spang

Subscribers: jenkins, khamidou, juan

Differential Revision: https://phab.nylas.com/D3099