Fix issue with labels not being deleted when they've been deleted from gmail
Summary:
Labels that are deleted from Gmail are never deleted in N1. There is a check to
make sure that we don't delete standard labels, but this checks that the canonical
name is None, while the default value for canonical names is actually an empty string.
Python doesn't evaluate '' to be None, so the condition is never satisfied and we
never delete any labels that aren't deleted from within N1. This diff changes the
condition to check if the canonical name is Falsy instead.

Test Plan: Tested locally

Reviewers: bengotow, annie, khamidou

Reviewed By: khamidou

Subscribers: khamidou, jenkins

Differential Revision: https://phab.nylas.com/D3126