Add batch tasks for syncback
Summary:
We want to amortize the cost of some IMAP operations (e.g. selecting a
folder) across a number of syncback actions. This diff adds a new
SyncbackBatchTask which is responsible for running a list of SyncbackTasks. The
next step will be to pull the CrispinClient up to the SyncbackBatchTask so we
can share it across SyncbackTasks.

Fixes T6979

Test Plan: Unit tests, run locally

Reviewers: drew, bengotow, khamidou, spang

Reviewed By: spang

Subscribers: jenkins

Maniphest Tasks: T6979

Differential Revision: https://phab.nylas.com/D3128