Switch to producer/consumer model for processing syncback actions
Summary:
Previously the syncback service was structured in such a way where we would
spawn a "worker" greenlet for processing an individual action. This was fine,
but it made it difficult to know whether we should be actively trying to find
more syncback actions to run or if all the workers are saturated and to just
wait until they're done.

This diff restructures things so that what was formerly a "worker" is now a
"task". We spawn a fixed number of consumers (now more accurately called
"workers"). The SyncbackService is now the "producer" which constantly polls
for new actions until all the workers are saturated. This allows us to eagerly
start looking for new actions as soon as a worker has finished processing its
current action.

Fixes T6973

Test Plan: Unit tests

Reviewers: spang, drew, bengotow

Reviewed By: bengotow

Subscribers: khamidou, jenkins

Maniphest Tasks: T6973

Differential Revision: https://phab.nylas.com/D3124