`FORCE INDEX` more conservatively
Summary: This is a rework of D3118. The previous fix got reverted because using the same index in all cases isn't right (for example if you're filtering over the messages in a given thread). This diff limits forcing the index to the cases where we're simply querying `/messages` or asking for messages before/after a given date.

Test Plan: Tested manually with a couple example queries. Checked that the generated SQL made sense and ran a couple queries in the MySQL optimizer to see which index would get picked up.

Reviewers: spang, rene

Reviewed By: rene

Subscribers: jenkins

Maniphest Tasks: T6967

Differential Revision: https://phab.nylas.com/D3132