Reduce amount of test data generated by hypothesis (#334)
We use hypothesis to autogenerate random test data for some of our IMAP sync
tests. The generation of this data is actually quite expensive due to the
amount of data we were generating. It also causes our tests to run slower since
we have so much data to process. This diff cuts down the amount of data we're
generating by around an order of magnitude.