Update folder roles if they've been changed (#370)
* Update the role for a Folder/Category if the role has changed

We've implemented code to guess folder roles, but these guesses aren't saved after the
initial creation of the folder. This diff changes that, so that users with accounts from
before the role guessing implementation can have their folders updated properly.

* Add some more folder edge cases

* Define getter/setter for Folder.canonical_name

The setter automatically updates the role for the corresponding Category, so that
people don't have to remember to update both.