Fixes for API error logging
The original new API for log_exception() had an optional message, but this was
removed before the patch was landed. Fix err() function to use the latest API
instead of passing in a message that was interpreted as the send_to_sentry
value.

Also, don't add 'error' keys to the request log context, which causes
log_exception() to not log additional information.

Lastly, don't send validation errors to err(), which will send them to
Sentry... raise an InputError instead and let the regular validation
error handler handle them.